@@ -30,6 +30,8 @@
   },
   "dependencies": {
     "@aws-sdk/credential-providers": "3.750.0",
+    "@grpc/grpc-js": "^1.13.2",
+    "@grpc/proto-loader": "^0.7.13",
     "@usebruno/common": "0.1.0",
     "@usebruno/converters": "^0.1.0",
     "@usebruno/filestore": "^0.1.0",
@@ -65,6 +67,7 @@
     "nanoid": "3.3.8",
     "qs": "^6.11.0",
     "socks-proxy-agent": "^8.0.2",
+    "tough-cookie": "^6.0.0",
     "uuid": "^9.0.0",
     "yup": "^0.32.11"
   },
@@ -72,7 +75,7 @@
     "dmg-license": "^1.0.11"
   },
   "devDependencies": {
-    "electron": "33.2.1",
+    "electron": "~37.2.6",
     "electron-builder": "25.1.8",
     "electron-devtools-installer": "^4.0.0"
   }
