@@ -7,6 +7,7 @@ import General from './General';
 import Proxy from './ProxySettings';
 import Display from './Display';
 import Keybindings from './Keybindings';
+import Beta from './Beta';
 
 import StyledWrapper from './StyledWrapper';
 
@@ -37,6 +38,10 @@ const Preferences = ({ onClose }) => {
         return <Keybindings close={onClose} />;
       }
 
+      case 'beta': {
+        return <Beta close={onClose} />;
+      }
+
       case 'support': {
         return <Support />;
       }
@@ -46,7 +51,7 @@ const Preferences = ({ onClose }) => {
   return (
     <StyledWrapper>
       <Modal size="lg" title="Preferences" handleCancel={onClose} hideFooter={true}>
-        <div className='flex flex-row gap-2 mx-[-1rem] !my-[-1.5rem] py-2'>
+        <div className="flex flex-row gap-2 mx-[-1rem] !my-[-1.5rem] py-2">
           <div className="flex flex-col items-center tabs" role="tablist">
             <div className={getTabClassname('general')} role="tab" onClick={() => setTab('general')}>
               General
@@ -62,6 +67,9 @@ const Preferences = ({ onClose }) => {
             </div>
             <div className={getTabClassname('support')} role="tab" onClick={() => setTab('support')}>
               Support
+            </div>
+            <div className={getTabClassname('beta')} role="tab" onClick={() => setTab('beta')}>
+              Beta
             </div>
           </div>
           <section className="flex flex-grow px-2 pt-2 pb-6 tab-panel">{getTabPanel(tab)}</section>
