@@ -71,81 +71,81 @@ const RequestBodyMode = ({ item, collection }) => {
     <StyledWrapper>
       <div className="inline-flex items-center cursor-pointer body-mode-selector">
         <Dropdown onCreate={onDropdownCreate} icon={<Icon />} placement="bottom-end">
-          <div className="label-item font-medium">Form</div>
-          <div
-            className="dropdown-item"
-            onClick={() => {
-              dropdownTippyRef.current.hide();
-              onModeChange('multipartForm');
-            }}
-          >
-            Multipart Form
-          </div>
-          <div
-            className="dropdown-item"
-            onClick={() => {
-              dropdownTippyRef.current.hide();
-              onModeChange('formUrlEncoded');
-            }}
-          >
-            Form URL Encoded
-          </div>
-          <div className="label-item font-medium">Raw</div>
-          <div
-            className="dropdown-item"
-            onClick={() => {
-              dropdownTippyRef.current.hide();
-              onModeChange('json');
-            }}
-          >
-            JSON
-          </div>
-          <div
-            className="dropdown-item"
-            onClick={() => {
-              dropdownTippyRef.current.hide();
-              onModeChange('xml');
-            }}
-          >
-            XML
-          </div>
-          <div
-            className="dropdown-item"
-            onClick={() => {
-              dropdownTippyRef.current.hide();
-              onModeChange('text');
-            }}
-          >
-            TEXT
-          </div>
-          <div
-            className="dropdown-item"
-            onClick={() => {
-              dropdownTippyRef.current.hide();
-              onModeChange('sparql');
-            }}
-          >
-            SPARQL
-          </div>
-          <div className="label-item font-medium">Other</div>
-          <div
-            className="dropdown-item"
-            onClick={() => {
-              dropdownTippyRef.current.hide();
-              onModeChange('file');
-            }}
-          >
-            File / Binary
-          </div>
-          <div
-            className="dropdown-item"
-            onClick={() => {
-              dropdownTippyRef.current.hide();
-              onModeChange('none');
-            }}
-          >
-            No Body
-          </div>
+        <div className="label-item font-medium">Form</div>
+        <div
+          className="dropdown-item"
+          onClick={() => {
+            dropdownTippyRef.current.hide();
+            onModeChange('multipartForm');
+          }}
+        >
+          Multipart Form
+        </div>
+        <div
+          className="dropdown-item"
+          onClick={() => {
+            dropdownTippyRef.current.hide();
+            onModeChange('formUrlEncoded');
+          }}
+        >
+          Form URL Encoded
+        </div>
+        <div className="label-item font-medium">Raw</div>
+        <div
+          className="dropdown-item"
+          onClick={() => {
+            dropdownTippyRef.current.hide();
+            onModeChange('json');
+          }}
+        >
+          JSON
+        </div>
+        <div
+          className="dropdown-item"
+          onClick={() => {
+            dropdownTippyRef.current.hide();
+            onModeChange('xml');
+          }}
+        >
+          XML
+        </div>
+        <div
+          className="dropdown-item"
+          onClick={() => {
+            dropdownTippyRef.current.hide();
+            onModeChange('text');
+          }}
+        >
+          TEXT
+        </div>
+        <div
+          className="dropdown-item"
+          onClick={() => {
+            dropdownTippyRef.current.hide();
+            onModeChange('sparql');
+          }}
+        >
+          SPARQL
+        </div>
+        <div className="label-item font-medium">Other</div>
+        <div
+          className="dropdown-item"
+          onClick={() => {
+            dropdownTippyRef.current.hide();
+            onModeChange('file');
+          }}
+        >
+          File / Binary
+        </div>
+        <div
+          className="dropdown-item"
+          onClick={() => {
+            dropdownTippyRef.current.hide();
+            onModeChange('none');
+          }}
+        >
+          No Body
+        </div>
         </Dropdown>
       </div>
       {(bodyMode === 'json' || bodyMode === 'xml') && (
