@@ -1,9 +1,9 @@
 import React from 'react';
 import StyledWrapper from './StyledWrapper';
+import isNumber from 'lodash/isNumber';
 
 const ResponseTime = ({ duration }) => {
   let durationToDisplay = '';
-
   if (duration > 1000) {
     // duration greater than a second
     let seconds = Math.floor(duration / 1000);
@@ -11,6 +11,10 @@ const ResponseTime = ({ duration }) => {
     durationToDisplay = seconds + '.' + decimal.toFixed(0) + 's';
   } else {
     durationToDisplay = duration + 'ms';
+  }
+
+  if (!isNumber(duration)) {
+    return null;
   }
 
   return <StyledWrapper className="ml-4">{durationToDisplay}</StyledWrapper>;
