@@ -3,9 +3,11 @@ import classnames from 'classnames';
 import StyledWrapper from './StyledWrapper';
 
 const RequestMethod = ({ item }) => {
-  if (!['http-request', 'graphql-request'].includes(item.type)) {
+  if (!['http-request', 'graphql-request', 'grpc-request'].includes(item.type)) {
     return null;
   }
+
+  const isGrpc = item.type === 'grpc-request';
 
   const getClassname = (method = '') => {
     method = method.toLocaleLowerCase();
@@ -16,7 +18,8 @@ const RequestMethod = ({ item }) => {
       'method-delete': method === 'delete',
       'method-patch': method === 'patch',
       'method-head': method === 'head',
-      'method-options': method == 'options'
+      'method-options': method === 'options',
+      'method-grpc': isGrpc,
     });
   };
 
@@ -24,7 +27,7 @@ const RequestMethod = ({ item }) => {
     <StyledWrapper>
       <div className={getClassname(item.request.method)}>
         <span className="uppercase">
-          {item.request.method.length > 5 ? item.request.method.substring(0, 3) : item.request.method}
+          {isGrpc ? 'grpc' : item.request.method.length > 5 ? item.request.method.substring(0, 3) : item.request.method}
         </span>
       </div>
     </StyledWrapper>
