@@ -5,6 +5,7 @@ import { refreshScreenWidth } from 'providers/ReduxStore/slices/app';
 import ConfirmAppClose from './ConfirmAppClose';
 import useIpcEvents from './useIpcEvents';
 import useTelemetry from './useTelemetry';
+import useGrpcEventListeners from 'utils/network/grpc-event-listeners';
 import StyledWrapper from './StyledWrapper';
 import { version } from '../../../package.json';
 
@@ -13,7 +14,7 @@ export const AppContext = React.createContext();
 export const AppProvider = (props) => {
   useTelemetry({ version });
   useIpcEvents();
-
+  useGrpcEventListeners();
   const dispatch = useDispatch();
 
   useEffect(() => {
