@@ -20,7 +20,7 @@ import {
   transformRequestToSaveToFilesystem
 } from 'utils/collections';
 import { uuid, waitForNextTick } from 'utils/common';
-import { cancelNetworkRequest, sendNetworkRequest } from 'utils/network';
+import { cancelNetworkRequest, sendGrpcRequest, sendNetworkRequest } from 'utils/network/index';
 import { callIpc } from 'utils/common/ipc';
 
 import {
@@ -39,7 +39,9 @@ import {
   collectionAddOauth2CredentialsByUrl,
   collectionClearOauth2CredentialsByUrl,
   initRunRequestEvent,
-  updateRunnerConfiguration as _updateRunnerConfiguration
+  updateRunnerConfiguration as _updateRunnerConfiguration,
+  updateActiveConnections,
+  saveRequest as _saveRequest
 } from './index';
 
 import { each } from 'lodash';
@@ -47,9 +49,19 @@ import { closeAllCollectionTabs } from 'providers/ReduxStore/slices/tabs';
 import { resolveRequestFilename } from 'utils/common/platform';
 import { parsePathParams, splitOnFirst } from 'utils/url/index';
 import { sendCollectionOauth2Request as _sendCollectionOauth2Request } from 'utils/network/index';
-import { getGlobalEnvironmentVariables, findCollectionByPathname, findEnvironmentInCollectionByName, getReorderedItemsInTargetDirectory, resetSequencesInFolder, getReorderedItemsInSourceDirectory, calculateDraggedItemNewPathname } from 'utils/collections/index';
+import {
+  getGlobalEnvironmentVariables,
+  findCollectionByPathname,
+  findEnvironmentInCollectionByName,
+  getReorderedItemsInTargetDirectory,
+  resetSequencesInFolder,
+  getReorderedItemsInSourceDirectory,
+  calculateDraggedItemNewPathname
+} from 'utils/collections/index';
 import { sanitizeName } from 'utils/common/regex';
 import { safeParseJSON, safeStringifyJSON } from 'utils/common/index';
+import { addTab } from 'providers/ReduxStore/slices/tabs';
+import { updateSettingsSelectedTab } from './index';
 
 export const renameCollection = (newName, collectionUid) => (dispatch, getState) => {
   const state = getState();
@@ -89,10 +101,16 @@ export const saveRequest = (itemUid, collectionUid, saveSilently) => (dispatch, 
         if (!saveSilently) {
           toast.success('Request saved successfully');
         }
+        dispatch(
+          _saveRequest({
+            itemUid,
+            collectionUid
+          })
+        );
       })
       .then(resolve)
       .catch((err) => {
-        toast.error('Failed to save request!');
+        toast.error(err.message || 'Failed to save request!');
         reject(err);
       });
   });
@@ -187,7 +205,7 @@ export const saveFolderRoot = (collectionUid, folderUid) => (dispatch, getState)
 
 export const sendCollectionOauth2Request = (collectionUid, itemUid) => (dispatch, getState) => {
   const state = getState();
-  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;  
+  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;
   const collection = findCollectionByUid(state.collections.collections, collectionUid);
 
   return new Promise((resolve, reject) => {
@@ -198,7 +216,10 @@ export const sendCollectionOauth2Request = (collectionUid, itemUid) => (dispatch
     let collectionCopy = cloneDeep(collection);
 
     // add selected global env variables to the collection object
-    const globalEnvironmentVariables = getGlobalEnvironmentVariables({ globalEnvironments, activeGlobalEnvironmentUid });
+    const globalEnvironmentVariables = getGlobalEnvironmentVariables({
+      globalEnvironments,
+      activeGlobalEnvironmentUid
+    });
     collectionCopy.globalEnvironmentVariables = globalEnvironmentVariables;
 
     const environment = findEnvironmentInCollection(collectionCopy, collection.activeEnvironmentUid);
@@ -221,7 +242,7 @@ export const sendCollectionOauth2Request = (collectionUid, itemUid) => (dispatch
 
 export const sendRequest = (item, collectionUid) => (dispatch, getState) => {
   const state = getState();
-  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;  
+  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;
   const collection = findCollectionByUid(state.collections.collections, collectionUid);
   const itemUid = item?.uid;
 
@@ -229,7 +250,7 @@ export const sendRequest = (item, collectionUid) => (dispatch, getState) => {
     if (!collection) {
       return reject(new Error('Collection not found'));
     }
-    
+
     let collectionCopy = cloneDeep(collection);
 
     const itemCopy = cloneDeep(item);
@@ -237,66 +258,79 @@ export const sendRequest = (item, collectionUid) => (dispatch, getState) => {
     const requestUid = uuid();
     itemCopy.requestUid = requestUid;
 
-    await dispatch(initRunRequestEvent({
-      requestUid,
-      itemUid,
-      collectionUid
-    }));
+    await dispatch(
+      initRunRequestEvent({
+        requestUid,
+        itemUid,
+        collectionUid
+      })
+    );
 
     // add selected global env variables to the collection object
-    const globalEnvironmentVariables = getGlobalEnvironmentVariables({ globalEnvironments, activeGlobalEnvironmentUid });
+    const globalEnvironmentVariables = getGlobalEnvironmentVariables({
+      globalEnvironments,
+      activeGlobalEnvironmentUid
+    });
     collectionCopy.globalEnvironmentVariables = globalEnvironmentVariables;
 
     const environment = findEnvironmentInCollection(collectionCopy, collectionCopy.activeEnvironmentUid);
-    sendNetworkRequest(itemCopy, collectionCopy, environment, collectionCopy.runtimeVariables)
-      .then((response) => {
-        // Ensure any timestamps in the response are converted to numbers
-        const serializedResponse = {
-          ...response,
-          timeline: response.timeline?.map(entry => ({
-            ...entry,
-            timestamp: entry.timestamp instanceof Date ? entry.timestamp.getTime() : entry.timestamp
-          }))
-        };
+    const isGrpcRequest = itemCopy.type === 'grpc-request';
+    if (isGrpcRequest) {
+      sendGrpcRequest(itemCopy, collectionCopy, environment, collectionCopy.runtimeVariables)
+        .then(resolve)
+        .catch((err) => {
+          toast.error(err.message);
+        });
+    } else {
+      sendNetworkRequest(itemCopy, collectionCopy, environment, collectionCopy.runtimeVariables)
+        .then((response) => {
+          // Ensure any timestamps in the response are converted to numbers
+          const serializedResponse = {
+            ...response,
+            timeline: response.timeline?.map((entry) => ({
+              ...entry,
+              timestamp: entry.timestamp instanceof Date ? entry.timestamp.getTime() : entry.timestamp
+            }))
+          };
 
-        return dispatch(
-          responseReceived({
-            itemUid,
-            collectionUid,
-            response: serializedResponse
-          })
-        );
-      })
-      .then(resolve)
-      .catch((err) => {
-        if (err && err.message === "Error invoking remote method 'send-http-request': Error: Request cancelled") {
-          console.log('>> request cancelled');
+          return dispatch(
+            responseReceived({
+              itemUid,
+              collectionUid,
+              response: serializedResponse
+            })
+          );
+        })
+        .then(resolve)
+        .catch((err) => {
+          if (err && err.message === "Error invoking remote method 'send-http-request': Error: Request cancelled") {
+            dispatch(
+              responseReceived({
+                itemUid,
+                collectionUid,
+                response: null
+              })
+            );
+            return;
+          }
+
+          const errorResponse = {
+            status: 'Error',
+            isError: true,
+            error: err.message ?? 'Something went wrong',
+            size: 0,
+            duration: 0
+          };
+
           dispatch(
             responseReceived({
               itemUid,
               collectionUid,
-              response: null
+              response: errorResponse
             })
           );
-          return;
-        }
-
-        const errorResponse = {
-          status: 'Error',
-          isError: true,
-          error: err.message ?? 'Something went wrong',
-          size: 0,
-          duration: 0
-        };
-
-        dispatch(
-          responseReceived({
-            itemUid,
-            collectionUid,
-            response: errorResponse
-          })
-        );
-      });
+        });
+    }
   });
 };
 
@@ -317,75 +351,79 @@ export const cancelRunnerExecution = (cancelTokenUid) => (dispatch) => {
   cancelNetworkRequest(cancelTokenUid).catch((err) => console.log(err));
 };
 
-export const runCollectionFolder = (collectionUid, folderUid, recursive, delay, tags, selectedRequestUids) => (dispatch, getState) => {
-  const state = getState();
-  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;
-  const collection = findCollectionByUid(state.collections.collections, collectionUid);
+export const runCollectionFolder =
+  (collectionUid, folderUid, recursive, delay, tags, selectedRequestUids) => (dispatch, getState) => {
+    const state = getState();
+    const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;
+    const collection = findCollectionByUid(state.collections.collections, collectionUid);
 
-  return new Promise((resolve, reject) => {
-    if (!collection) {
-      return reject(new Error('Collection not found'));
-    }
-
-    let collectionCopy = cloneDeep(collection);
-
-    // add selected global env variables to the collection object
-    const globalEnvironmentVariables = getGlobalEnvironmentVariables({ globalEnvironments, activeGlobalEnvironmentUid });
-    collectionCopy.globalEnvironmentVariables = globalEnvironmentVariables;
-
-    const folder = findItemInCollection(collectionCopy, folderUid);
-
-    if (folderUid && !folder) {
-      return reject(new Error('Folder not found'));
-    }
-
-    const environment = findEnvironmentInCollection(collectionCopy, collection.activeEnvironmentUid);
-
-    dispatch(
-      resetRunResults({
-        collectionUid: collection.uid
-      })
-    );
-
-    // to only include those requests in the specified order while preserving folder data
-    if (selectedRequestUids && selectedRequestUids.length > 0) {
-      const newItems = [];
-      
-      selectedRequestUids.forEach((uid, index) => {
-        const requestItem = findItemInCollection(collectionCopy, uid);
-        if (requestItem) {
-          const clonedRequest = cloneDeep(requestItem);
-          clonedRequest.seq = index + 1;
-          newItems.push(clonedRequest);
-        }
-      });
-      
-      if (folder) {
-        folder.items = newItems;
-      } else {
-        collectionCopy.items = newItems;
+    return new Promise((resolve, reject) => {
+      if (!collection) {
+        return reject(new Error('Collection not found'));
       }
-    }
 
-    const { ipcRenderer } = window;
-    ipcRenderer
-      .invoke(
-        'renderer:run-collection-folder',
-        folder,
-        collectionCopy,
-        environment,
-        collectionCopy.runtimeVariables,
-        recursive,
-        delay,
-        tags
-      )
-      .then(resolve)
-      .catch((err) => {
-        toast.error(get(err, 'error.message') || 'Something went wrong!');
-        reject(err);
+      let collectionCopy = cloneDeep(collection);
+
+      // add selected global env variables to the collection object
+      const globalEnvironmentVariables = getGlobalEnvironmentVariables({
+        globalEnvironments,
+        activeGlobalEnvironmentUid
       });
-  });
-};
+      collectionCopy.globalEnvironmentVariables = globalEnvironmentVariables;
+
+      const folder = findItemInCollection(collectionCopy, folderUid);
+
+      if (folderUid && !folder) {
+        return reject(new Error('Folder not found'));
+      }
+
+      const environment = findEnvironmentInCollection(collectionCopy, collection.activeEnvironmentUid);
+
+      dispatch(
+        resetRunResults({
+          collectionUid: collection.uid
+        })
+      );
+
+      // to only include those requests in the specified order while preserving folder data
+      if (selectedRequestUids && selectedRequestUids.length > 0) {
+        const newItems = [];
+
+        selectedRequestUids.forEach((uid, index) => {
+          const requestItem = findItemInCollection(collectionCopy, uid);
+          if (requestItem) {
+            const clonedRequest = cloneDeep(requestItem);
+            clonedRequest.seq = index + 1;
+            newItems.push(clonedRequest);
+          }
+        });
+
+        if (folder) {
+          folder.items = newItems;
+        } else {
+          collectionCopy.items = newItems;
+        }
+      }
+
+      const { ipcRenderer } = window;
+      ipcRenderer
+        .invoke(
+          'renderer:run-collection-folder',
+          folder,
+          collectionCopy,
+          environment,
+          collectionCopy.runtimeVariables,
+          recursive,
+          delay,
+          tags
+        )
+        .then(resolve)
+        .catch((err) => {
+          toast.error(get(err, 'error.message') || 'Something went wrong!');
+          reject(err);
+        });
+    });
+  };
 
 export const newFolder = (folderName, directoryName, collectionUid, itemUid) => (dispatch, getState) => {
   const state = getState();
@@ -410,7 +448,7 @@ export const newFolder = (folderName, directoryName, collectionUid, itemUid) => 
         const folderBruJsonData = {
           meta: {
             name: folderName,
-            seq: items?.length + 1 
+            seq: items?.length + 1
           },
           request: {
             auth: {
@@ -424,7 +462,7 @@ export const newFolder = (folderName, directoryName, collectionUid, itemUid) => 
           .then(resolve)
           .catch((error) => {
             toast.error('Failed to create a new folder!');
-            reject(error)
+            reject(error);
           });
       } else {
         return reject(new Error('Duplicate folder names under same parent folder are not allowed'));
@@ -443,7 +481,7 @@ export const newFolder = (folderName, directoryName, collectionUid, itemUid) => 
           const folderBruJsonData = {
             meta: {
               name: folderName,
-              seq: items?.length + 1 
+              seq: items?.length + 1
             },
             request: {
               auth: {
@@ -457,7 +495,7 @@ export const newFolder = (folderName, directoryName, collectionUid, itemUid) => 
             .then(resolve)
             .catch((error) => {
               toast.error('Failed to create a new folder!');
-              reject(error)
+              reject(error);
             });
         } else {
           return reject(new Error('Duplicate folder names under same parent folder are not allowed'));
@@ -469,66 +507,68 @@ export const newFolder = (folderName, directoryName, collectionUid, itemUid) => 
   });
 };
 
-export const renameItem = ({ newName, newFilename, itemUid, collectionUid }) => (dispatch, getState) => {
-  const state = getState();
-  const collection = findCollectionByUid(state.collections.collections, collectionUid);
+export const renameItem =
+  ({ newName, newFilename, itemUid, collectionUid }) =>
+  (dispatch, getState) => {
+    const state = getState();
+    const collection = findCollectionByUid(state.collections.collections, collectionUid);
 
-  return new Promise((resolve, reject) => {
-    if (!collection) {
-      return reject(new Error('Collection not found'));
-    }
+    return new Promise((resolve, reject) => {
+      if (!collection) {
+        return reject(new Error('Collection not found'));
+      }
 
-    const collectionCopy = cloneDeep(collection);
-    const item = findItemInCollection(collectionCopy, itemUid);
-    if (!item) {
-      return reject(new Error('Unable to locate item'));
-    }
+      const collectionCopy = cloneDeep(collection);
+      const item = findItemInCollection(collectionCopy, itemUid);
+      if (!item) {
+        return reject(new Error('Unable to locate item'));
+      }
 
-    const { ipcRenderer } = window;
+      const { ipcRenderer } = window;
 
-    const renameName = async () => {
-      return ipcRenderer.invoke('renderer:rename-item-name', { itemPath: item.pathname, newName })
-        .catch((err) => {
+      const renameName = async () => {
+        return ipcRenderer.invoke('renderer:rename-item-name', { itemPath: item.pathname, newName }).catch((err) => {
           toast.error('Failed to rename the item name');
           console.error(err);
           throw new Error('Failed to rename the item name');
         });
-    };
+      };
 
-    const renameFile = async () => {
-      const dirname = path.dirname(item.pathname);
-      let newPath = '';
-      if (item.type === 'folder') {
-        newPath = path.join(dirname, trim(newFilename));
-      } else {
-        const filename = resolveRequestFilename(newFilename);
-        newPath = path.join(dirname, filename);
+      const renameFile = async () => {
+        const dirname = path.dirname(item.pathname);
+        let newPath = '';
+        if (item.type === 'folder') {
+          newPath = path.join(dirname, trim(newFilename));
+        } else {
+          const filename = resolveRequestFilename(newFilename);
+          newPath = path.join(dirname, filename);
+        }
+
+        return ipcRenderer
+          .invoke('renderer:rename-item-filename', { oldPath: item.pathname, newPath, newName, newFilename })
+          .catch((err) => {
+            toast.error('Failed to rename the file');
+            console.error(err);
+            throw new Error('Failed to rename the file');
+          });
+      };
+
+      let renameOperation = null;
+      if (newName) renameOperation = renameName;
+      if (newFilename) renameOperation = renameFile;
+
+      if (!renameOperation) {
+        resolve();
       }
 
-      return ipcRenderer.invoke('renderer:rename-item-filename', { oldPath: item.pathname, newPath, newName, newFilename })
-        .catch((err) => {
-          toast.error('Failed to rename the file');
-          console.error(err);
-          throw new Error('Failed to rename the file');
-        });
-    };
-
-    let renameOperation = null;
-    if (newName) renameOperation = renameName;
-    if (newFilename) renameOperation = renameFile;
-
-    if (!renameOperation) {
-      resolve();
-    }
-    
-    renameOperation()
-      .then(() => {
-        toast.success('Item renamed successfully');
-        resolve();
-      })
-      .catch((err) => reject(err));
-  });
-};
+      renameOperation()
+        .then(() => {
+          toast.success('Item renamed successfully');
+          resolve();
+        })
+        .catch((err) => reject(err));
+    });
+  };
 
 export const cloneItem = (newName, newFilename, itemUid, collectionUid) => (dispatch, getState) => {
   const state = getState();
@@ -560,9 +600,9 @@ export const cloneItem = (newName, newFilename, itemUid, collectionUid) => (disp
       set(item, 'filename', newFilename);
       set(item, 'root.meta.name', newName);
       set(item, 'root.meta.seq', parentFolder?.items?.length + 1);
-      
+
       const collectionPath = path.join(parentFolder.pathname, newFilename);
-      
+
       const { ipcRenderer } = window;
       ipcRenderer.invoke('renderer:clone-folder', item, collectionPath).then(resolve).catch(reject);
       return;
@@ -662,116 +702,155 @@ export const sortCollections = (payload) => (dispatch) => {
   dispatch(_sortCollections(payload));
 };
 
-export const moveItem = ({ targetDirname, sourcePathname }) => (dispatch, getState) => {
-  return new Promise((resolve, reject) => {
-    const { ipcRenderer } = window;
+export const moveItem =
+  ({ targetDirname, sourcePathname }) =>
+  (dispatch, getState) => {
+    return new Promise((resolve, reject) => {
+      const { ipcRenderer } = window;
 
-    ipcRenderer.invoke('renderer:move-item', { targetDirname, sourcePathname })
-      .then(resolve)
-      .catch(reject);
-  });
-}
+      ipcRenderer.invoke('renderer:move-item', { targetDirname, sourcePathname }).then(resolve).catch(reject);
+    });
+  };
 
-export const handleCollectionItemDrop = ({ targetItem, draggedItem, dropType, collectionUid }) => (dispatch, getState) => {
-  const state = getState();
-  const collection = findCollectionByUid(state.collections.collections, collectionUid);
-  const { uid: draggedItemUid, pathname: draggedItemPathname } = draggedItem;
-  const { uid: targetItemUid, pathname: targetItemPathname } = targetItem;
-  const targetItemDirectory = findParentItemInCollection(collection, targetItemUid) || collection;
-  const targetItemDirectoryItems = cloneDeep(targetItemDirectory.items);
-  const draggedItemDirectory = findParentItemInCollection(collection, draggedItemUid) || collection;
-  const draggedItemDirectoryItems = cloneDeep(draggedItemDirectory.items);
+export const handleCollectionItemDrop =
+  ({ targetItem, draggedItem, dropType, collectionUid }) =>
+  (dispatch, getState) => {
+    const state = getState();
+    const collection = findCollectionByUid(state.collections.collections, collectionUid);
+    const { uid: draggedItemUid, pathname: draggedItemPathname } = draggedItem;
+    const { uid: targetItemUid, pathname: targetItemPathname } = targetItem;
+    const targetItemDirectory = findParentItemInCollection(collection, targetItemUid) || collection;
+    const targetItemDirectoryItems = cloneDeep(targetItemDirectory.items);
+    const draggedItemDirectory = findParentItemInCollection(collection, draggedItemUid) || collection;
+    const draggedItemDirectoryItems = cloneDeep(draggedItemDirectory.items);
 
-  const handleMoveToNewLocation = async ({ draggedItem, draggedItemDirectoryItems, targetItem, targetItemDirectoryItems, newPathname, dropType }) => {
-    const { uid: targetItemUid } = targetItem;
-    const { pathname: draggedItemPathname, uid: draggedItemUid } = draggedItem;
-    
-    const newDirname = path.dirname(newPathname);
-    await dispatch(moveItem({
-      targetDirname: newDirname,
-      sourcePathname: draggedItemPathname
-    }));
+    const handleMoveToNewLocation = async ({
+      draggedItem,
+      draggedItemDirectoryItems,
+      targetItem,
+      targetItemDirectoryItems,
+      newPathname,
+      dropType
+    }) => {
+      const { uid: targetItemUid } = targetItem;
+      const { pathname: draggedItemPathname, uid: draggedItemUid } = draggedItem;
 
-    // Update sequences in the source directory
-    if (draggedItemDirectoryItems?.length) {
-      // reorder items in the source directory
-      const draggedItemDirectoryItemsWithoutDraggedItem = draggedItemDirectoryItems.filter(i => i.uid !== draggedItemUid);
-      const reorderedSourceItems = getReorderedItemsInSourceDirectory({ items: draggedItemDirectoryItemsWithoutDraggedItem });
-      if (reorderedSourceItems?.length) {
-        await dispatch(updateItemsSequences({ itemsToResequence: reorderedSourceItems }));
+      const newDirname = path.dirname(newPathname);
+      await dispatch(
+        moveItem({
+          targetDirname: newDirname,
+          sourcePathname: draggedItemPathname
+        })
+      );
+
+      // Update sequences in the source directory
+      if (draggedItemDirectoryItems?.length) {
+        // reorder items in the source directory
+        const draggedItemDirectoryItemsWithoutDraggedItem = draggedItemDirectoryItems.filter(
+          (i) => i.uid !== draggedItemUid
+        );
+        const reorderedSourceItems = getReorderedItemsInSourceDirectory({
+          items: draggedItemDirectoryItemsWithoutDraggedItem
+        });
+        if (reorderedSourceItems?.length) {
+          await dispatch(updateItemsSequences({ itemsToResequence: reorderedSourceItems }));
+        }
       }
-    }
 
-    // Update sequences in the target directory (if dropping adjacent)
-    if (dropType === 'adjacent') {
-      const targetItemSequence = targetItemDirectoryItems.findIndex(i => i.uid === targetItemUid)?.seq;
+      // Update sequences in the target directory (if dropping adjacent)
+      if (dropType === 'adjacent') {
+        const targetItemSequence = targetItemDirectoryItems.findIndex((i) => i.uid === targetItemUid)?.seq;
 
-      const draggedItemWithNewPathAndSequence = {
-        ...draggedItem,
-        pathname: newPathname,
-        seq: targetItemSequence
-      };
+        const draggedItemWithNewPathAndSequence = {
+          ...draggedItem,
+          pathname: newPathname,
+          seq: targetItemSequence
+        };
 
-      // draggedItem is added to the targetItem's directory
-      const reorderedTargetItems = getReorderedItemsInTargetDirectory({
-        items: [ ...targetItemDirectoryItems, draggedItemWithNewPathAndSequence ],
+        // draggedItem is added to the targetItem's directory
+        const reorderedTargetItems = getReorderedItemsInTargetDirectory({
+          items: [...targetItemDirectoryItems, draggedItemWithNewPathAndSequence],
+          targetItemUid,
+          draggedItemUid
+        });
+
+        if (reorderedTargetItems?.length) {
+          await dispatch(updateItemsSequences({ itemsToResequence: reorderedTargetItems }));
+        }
+      }
+    };
+
+    const handleReorderInSameLocation = async ({ draggedItem, targetItem, targetItemDirectoryItems }) => {
+      const { uid: targetItemUid } = targetItem;
+      const { uid: draggedItemUid } = draggedItem;
+
+      // reorder items in the targetItem's directory
+      const reorderedItems = getReorderedItemsInTargetDirectory({
+        items: targetItemDirectoryItems,
         targetItemUid,
         draggedItemUid
       });
 
-      if (reorderedTargetItems?.length) {
-        await dispatch(updateItemsSequences({ itemsToResequence: reorderedTargetItems }));
+      if (reorderedItems?.length) {
+        await dispatch(updateItemsSequences({ itemsToResequence: reorderedItems }));
       }
-    }
-  };
+    };
 
-  const handleReorderInSameLocation = async ({ draggedItem, targetItem, targetItemDirectoryItems }) => {
-    const { uid: targetItemUid } = targetItem;
-    const { uid: draggedItemUid } = draggedItem;
-
-    // reorder items in the targetItem's directory
-    const reorderedItems = getReorderedItemsInTargetDirectory({
-      items: targetItemDirectoryItems,
-      targetItemUid,
-      draggedItemUid
+    return new Promise(async (resolve, reject) => {
+      try {
+        const newPathname = calculateDraggedItemNewPathname({
+          draggedItem,
+          targetItem,
+          dropType,
+          collectionPathname: collection.pathname
+        });
+        if (!newPathname) return;
+        if (targetItemPathname?.startsWith(draggedItemPathname)) return;
+        if (newPathname !== draggedItemPathname) {
+          await handleMoveToNewLocation({
+            targetItem,
+            targetItemDirectoryItems,
+            draggedItem,
+            draggedItemDirectoryItems,
+            newPathname,
+            dropType
+          });
+        } else {
+          await handleReorderInSameLocation({ draggedItem, targetItemDirectoryItems, targetItem });
+        }
+        resolve();
+      } catch (error) {
+        console.error(error);
+        toast.error(error?.message);
+        reject(error);
+      }
     });
-
-    if (reorderedItems?.length) {
-      await dispatch(updateItemsSequences({ itemsToResequence: reorderedItems }));
-    }
   };
 
-  return new Promise(async (resolve, reject) => {
-    try {
-      const newPathname = calculateDraggedItemNewPathname({ draggedItem, targetItem, dropType, collectionPathname: collection.pathname });
-      if (!newPathname) return;
-      if (targetItemPathname?.startsWith(draggedItemPathname)) return;
-      if (newPathname !== draggedItemPathname) {
-        await handleMoveToNewLocation({ targetItem, targetItemDirectoryItems, draggedItem, draggedItemDirectoryItems, newPathname, dropType });
-      } else {
-        await handleReorderInSameLocation({ draggedItem, targetItemDirectoryItems, targetItem });
-      }
-      resolve();
-    } catch (error) {
-      console.error(error);
-      toast.error(error?.message);
-      reject(error);
-    }
-  })
-}
+export const updateItemsSequences =
+  ({ itemsToResequence }) =>
+  (dispatch, getState) => {
+    return new Promise((resolve, reject) => {
+      const { ipcRenderer } = window;
 
-export const updateItemsSequences = ({ itemsToResequence }) => (dispatch, getState) => {
-  return new Promise((resolve, reject) => {
-    const { ipcRenderer } = window;
-
-    ipcRenderer.invoke('renderer:resequence-items', itemsToResequence)
-      .then(resolve)
-      .catch(reject);
-  });
-}
+      ipcRenderer.invoke('renderer:resequence-items', itemsToResequence).then(resolve).catch(reject);
+    });
+  };
 
 export const newHttpRequest = (params) => (dispatch, getState) => {
-  const { requestName, filename, requestType, requestUrl, requestMethod, collectionUid, itemUid, headers, body, auth, settings } = params;
+  const {
+    requestName,
+    filename,
+    requestType,
+    requestUrl,
+    requestMethod,
+    collectionUid,
+    itemUid,
+    headers,
+    body,
+    auth,
+    settings
+  } = params;
 
   return new Promise((resolve, reject) => {
     const state = getState();
@@ -838,18 +917,21 @@ export const newHttpRequest = (params) => (dispatch, getState) => {
         const fullName = path.join(collection.pathname, resolvedFilename);
         const { ipcRenderer } = window;
 
-        ipcRenderer.invoke('renderer:new-request', fullName, item).then(() => {
-          // task middleware will track this and open the new request in a new tab once request is created
-          dispatch(
-            insertTaskIntoQueue({
-              uid: uuid(),
-              type: 'OPEN_REQUEST',
-              collectionUid,
-              itemPathname: fullName
-            })
-          );
-          resolve();
-        }).catch(reject);
+        ipcRenderer
+          .invoke('renderer:new-request', fullName, item)
+          .then(() => {
+            // task middleware will track this and open the new request in a new tab once request is created
+            dispatch(
+              insertTaskIntoQueue({
+                uid: uuid(),
+                type: 'OPEN_REQUEST',
+                collectionUid,
+                itemPathname: fullName
+              })
+            );
+            resolve();
+          })
+          .catch(reject);
       } else {
         return reject(new Error('Duplicate request names are not allowed under the same folder'));
       }
@@ -865,23 +947,151 @@ export const newHttpRequest = (params) => (dispatch, getState) => {
         if (!reqWithSameNameExists) {
           const fullName = path.join(currentItem.pathname, resolvedFilename);
           const { ipcRenderer } = window;
-          ipcRenderer.invoke('renderer:new-request', fullName, item).then(() => {
-            // task middleware will track this and open the new request in a new tab once request is created
-            dispatch(
-              insertTaskIntoQueue({
-                uid: uuid(),
-                type: 'OPEN_REQUEST',
-                collectionUid,
-                itemPathname: fullName
-              })
-            );
-            resolve();
-          }).catch(reject);
+          ipcRenderer
+            .invoke('renderer:new-request', fullName, item)
+            .then(() => {
+              // task middleware will track this and open the new request in a new tab once request is created
+              dispatch(
+                insertTaskIntoQueue({
+                  uid: uuid(),
+                  type: 'OPEN_REQUEST',
+                  collectionUid,
+                  itemPathname: fullName
+                })
+              );
+              resolve();
+            })
+            .catch(reject);
         } else {
           return reject(new Error('Duplicate request names are not allowed under the same folder'));
         }
       }
     }
+  });
+};
+
+export const newGrpcRequest = (params) => (dispatch, getState) => {
+  const { requestName, filename, requestUrl, collectionUid, body, auth, headers, itemUid } = params;
+
+  return new Promise((resolve, reject) => {
+    const state = getState();
+    const collection = findCollectionByUid(state.collections.collections, collectionUid);
+    if (!collection) {
+      return reject(new Error('Collection not found'));
+    }
+    // do we need to handle query, path params for grpc requests?
+    // skipping for now
+
+    const item = {
+      uid: uuid(),
+      name: requestName,
+      filename,
+      type: 'grpc-request',
+      headers: headers ?? [],
+      request: {
+        url: requestUrl,
+        body: body ?? {
+          mode: 'grpc',
+          grpc: [
+            {
+              name: 'message 1',
+              content: '{}'
+            }
+          ]
+        },
+        auth: auth ?? {
+          mode: 'inherit'
+        }
+      }
+    };
+
+    const resolvedFilename = resolveRequestFilename(filename);
+    const fullName = path.join(collection.pathname, resolvedFilename);
+    const { ipcRenderer } = window;
+
+    // Set the seq field for gRPC requests
+    const items = filter(collection.items, (i) => isItemAFolder(i) || isItemARequest(i));
+    item.seq = items.length + 1;
+
+    ipcRenderer
+      .invoke('renderer:new-request', fullName, item)
+      .then(() => {
+        // task middleware will track this and open the new request in a new tab once request is created
+        dispatch(
+          insertTaskIntoQueue({
+            uid: uuid(),
+            type: 'OPEN_REQUEST',
+            collectionUid,
+            itemPathname: fullName
+          })
+        );
+        resolve();
+      })
+      .catch(reject);
+  });
+};
+
+export const loadGrpcMethodsFromReflection = (item, collectionUid, url) => async (dispatch, getState) => {
+  const state = getState();
+  const collection = findCollectionByUid(state.collections.collections, collectionUid);
+  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;
+
+  return new Promise((resolve, reject) => {
+    if (!collection) {
+      return reject(new Error('Collection not found'));
+    }
+
+    const itemCopy = cloneDeep(item);
+    const requestItem = itemCopy.draft ? itemCopy.draft : itemCopy;
+    requestItem.request.url = url;
+    const collectionCopy = cloneDeep(collection);
+    const globalEnvironmentVariables = getGlobalEnvironmentVariables({
+      globalEnvironments,
+      activeGlobalEnvironmentUid
+    });
+    collectionCopy.globalEnvironmentVariables = globalEnvironmentVariables;
+    const environment = findEnvironmentInCollection(collectionCopy, collectionCopy.activeEnvironmentUid);
+    const runtimeVariables = collectionCopy.runtimeVariables;
+
+    const { ipcRenderer } = window;
+    ipcRenderer
+      .invoke('grpc:load-methods-reflection', {
+        request: requestItem,
+        collection: collectionCopy,
+        environment,
+        runtimeVariables
+      })
+      .then(resolve)
+      .catch(reject);
+  });
+};
+
+export const generateGrpcurlCommand = (item, collectionUid) => async (dispatch, getState) => {
+  const state = getState();
+  const collection = findCollectionByUid(state.collections.collections, collectionUid);
+  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;
+
+  return new Promise((resolve, reject) => {
+    if (!collection) {
+      return reject(new Error('Collection not found'));
+    }
+
+    const itemCopy = cloneDeep(item);
+    const collectionCopy = cloneDeep(collection);
+
+    const globalEnvironmentVariables = getGlobalEnvironmentVariables({
+      globalEnvironments,
+      activeGlobalEnvironmentUid
+    });
+    collectionCopy.globalEnvironmentVariables = globalEnvironmentVariables;
+    const environment = findEnvironmentInCollection(collectionCopy, collectionCopy.activeEnvironmentUid);
+    const runtimeVariables = collectionCopy.runtimeVariables;
+
+    const { ipcRenderer } = window;
+    ipcRenderer
+      .invoke('grpc:generate-grpcurl', { request: itemCopy, collection: collectionCopy, environment, runtimeVariables })
+      .then(resolve)
+      .catch(reject);
   });
 };
 
@@ -919,7 +1129,7 @@ export const importEnvironment = (name, variables, collectionUid) => (dispatch, 
     if (!collection) {
       return reject(new Error('Collection not found'));
     }
-    
+
     const sanitizedName = sanitizeName(name);
 
     const { ipcRenderer } = window;
@@ -954,7 +1164,7 @@ export const copyEnvironment = (name, baseEnvUid, collectionUid) => (dispatch, g
       return reject(new Error('Environment not found'));
     }
 
-    const sanitizedName = sanitizeName(name); 
+    const sanitizedName = sanitizeName(name);
 
     const { ipcRenderer } = window;
     ipcRenderer
@@ -1050,6 +1260,68 @@ export const saveEnvironment = (variables, environmentUid, collectionUid) => (di
   });
 };
 
+export const mergeAndPersistEnvironment =
+  ({ persistentEnvVariables, collectionUid }) =>
+  (_dispatch, getState) => {
+    return new Promise((resolve, reject) => {
+      const state = getState();
+      const collection = findCollectionByUid(state.collections.collections, collectionUid);
+
+      if (!collection) {
+        return reject(new Error('Collection not found'));
+      }
+
+      const environmentUid = collection.activeEnvironmentUid;
+      if (!environmentUid) {
+        return reject(new Error('No active environment found'));
+      }
+
+      const collectionCopy = cloneDeep(collection);
+      const environment = findEnvironmentInCollection(collectionCopy, environmentUid);
+      if (!environment) {
+        return reject(new Error('Environment not found'));
+      }
+
+      // Only proceed if there are persistent variables to save
+      if (!persistentEnvVariables || Object.keys(persistentEnvVariables).length === 0) {
+        return resolve();
+      }
+
+      let existingVars = environment.variables || [];
+
+      let normalizedNewVars = Object.entries(persistentEnvVariables).map(([name, value]) => ({
+        uid: uuid(),
+        name,
+        value,
+        type: 'text',
+        enabled: true,
+        secret: false
+      }));
+
+      const merged = existingVars.map((v) => {
+        const found = normalizedNewVars.find((nv) => nv.name === v.name);
+        if (found) {
+          return { ...v, value: found.value };
+        }
+        return v;
+      });
+      normalizedNewVars.forEach((nv) => {
+        if (!merged.some((v) => v.name === nv.name)) {
+          merged.push(nv);
+        }
+      });
+
+      environment.variables = merged;
+
+      const { ipcRenderer } = window;
+      environmentSchema
+        .validate(environment)
+        .then(() => ipcRenderer.invoke('renderer:save-environment', collection.pathname, environment))
+        .then(resolve)
+        .catch(reject);
+    });
+  };
+
 export const selectEnvironment = (environmentUid, collectionUid) => (dispatch, getState) => {
   return new Promise((resolve, reject) => {
     const state = getState();
@@ -1060,16 +1332,17 @@ export const selectEnvironment = (environmentUid, collectionUid) => (dispatch, g
 
     const collectionCopy = cloneDeep(collection);
 
-    const environmentName = environmentUid 
-      ? findEnvironmentInCollection(collectionCopy, environmentUid)?.name 
-      : null;
+    const environmentName = environmentUid ? findEnvironmentInCollection(collectionCopy, environmentUid)?.name : null;
 
     if (environmentUid && !environmentName) {
       return reject(new Error('Environment not found'));
-    }  
+    }
 
     const { ipcRenderer } = window;
-    ipcRenderer.invoke('renderer:update-ui-state-snapshot', { type: 'COLLECTION_ENVIRONMENT', data: { collectionPath: collection?.pathname, environmentName }});
+    ipcRenderer.invoke('renderer:update-ui-state-snapshot', {
+      type: 'COLLECTION_ENVIRONMENT',
+      data: { collectionPath: collection?.pathname, environmentName }
+    });
 
     dispatch(_selectEnvironment({ environmentUid, collectionUid }));
     resolve();
@@ -1110,17 +1383,12 @@ export const browseDirectory = () => (dispatch, getState) => {
   });
 };
 
-export const browseFiles =
-  (filters, properties) =>
-  (_dispatch, _getState) => {
-    const { ipcRenderer } = window;
+export const browseFiles = (filters, properties) => (_dispatch, _getState) => {
+  const { ipcRenderer } = window;
 
-    return new Promise((resolve, reject) => {
-      ipcRenderer
-        .invoke('renderer:browse-files', filters, properties)
-        .then(resolve)
-        .catch(reject);
-    });
+  return new Promise((resolve, reject) => {
+    ipcRenderer.invoke('renderer:browse-files', filters, properties).then(resolve).catch(reject);
+  });
 };
 
 export const updateBrunoConfig = (brunoConfig, collectionUid) => (dispatch, getState) => {
@@ -1227,21 +1495,20 @@ export const importCollection = (collection, collectionLocation) => (dispatch, g
   });
 };
 
-export const moveCollectionAndPersist = ({ draggedItem, targetItem }) => (dispatch, getState) => {
-  dispatch(moveCollection({ draggedItem, targetItem }));
+export const moveCollectionAndPersist =
+  ({ draggedItem, targetItem }) =>
+  (dispatch, getState) => {
+    dispatch(moveCollection({ draggedItem, targetItem }));
 
-  return new Promise((resolve, reject) => {
-    const { ipcRenderer } = window;
-    const state = getState();
+    return new Promise((resolve, reject) => {
+      const { ipcRenderer } = window;
+      const state = getState();
 
-    const collectionPaths = state.collections.collections.map((collection) => collection.pathname);
+      const collectionPaths = state.collections.collections.map((collection) => collection.pathname);
 
-    ipcRenderer
-      .invoke('renderer:update-collection-paths', collectionPaths) 
-      .then(resolve)
-      .catch(reject);
-  });
-};
+      ipcRenderer.invoke('renderer:update-collection-paths', collectionPaths).then(resolve).catch(reject);
+    });
+  };
 
 export const saveCollectionSecurityConfig = (collectionUid, securityConfig) => (dispatch, getState) => {
   return new Promise((resolve, reject) => {
@@ -1259,13 +1526,12 @@ export const saveCollectionSecurityConfig = (collectionUid, securityConfig) => (
   });
 };
 
-
 export const hydrateCollectionWithUiStateSnapshot = (payload) => (dispatch, getState) => {
   const collectionSnapshotData = payload;
   return new Promise((resolve, reject) => {
     const state = getState();
     try {
-      if(!collectionSnapshotData) resolve();
+      if (!collectionSnapshotData) resolve();
       const { pathname, selectedEnvironment } = collectionSnapshotData;
       const collection = findCollectionByPathname(state.collections.collections, pathname);
       const collectionCopy = cloneDeep(collection);
@@ -1280,10 +1546,9 @@ export const hydrateCollectionWithUiStateSnapshot = (payload) => (dispatch, getS
       }
 
       // todo: add any other redux state that you want to save
-      
+
       resolve();
-    }
-    catch(error) {
+    } catch (error) {
       reject(error);
     }
   });
@@ -1324,7 +1589,7 @@ export const refreshOauth2Credentials = (payload) => async (dispatch, getState) 
   request.globalEnvironmentVariables = globalEnvironmentVariables;
   return new Promise((resolve, reject) => {
     window.ipcRenderer
-      .invoke('renderer:refresh-oauth2-credentials', { request, collection })
+      .invoke('renderer:refresh-oauth2-credentials', { itemUid, request, collection })
       .then(({ credentials, url, collectionUid, debugInfo, credentialsId }) => {
         dispatch(
           collectionAddOauth2CredentialsByUrl({
@@ -1352,7 +1617,7 @@ export const clearOauth2Cache = (payload) => async (dispatch, getState) => {
         dispatch(
           collectionClearOauth2CredentialsByUrl({
             url,
-            collectionUid,
+            collectionUid
           })
         );
         resolve();
@@ -1361,6 +1626,7 @@ export const clearOauth2Cache = (payload) => async (dispatch, getState) => {
   });
 };
 
+// todo: could be removed
 export const loadRequestViaWorker = ({ collectionUid, pathname }) => (dispatch, getState) => {
   return new Promise(async (resolve, reject) => {
     const { ipcRenderer } = window;
@@ -1368,6 +1634,7 @@ export const loadRequestViaWorker = ({ collectionUid, pathname }) => (dispatch, 
   });
 };
 
+// todo: could be removed
 export const loadRequest = ({ collectionUid, pathname }) => (dispatch, getState) => {
   return new Promise(async (resolve, reject) => {
     const { ipcRenderer } = window;
@@ -1375,30 +1642,76 @@ export const loadRequest = ({ collectionUid, pathname }) => (dispatch, getState)
   });
 };
 
-export const mountCollection = ({ collectionUid, collectionPathname, brunoConfig }) => (dispatch, getState) => {
-  dispatch(updateCollectionMountStatus({ collectionUid, mountStatus: 'mounting' }));
+export const loadLargeRequest = ({ collectionUid, pathname }) => (dispatch, getState) => {
   return new Promise(async (resolve, reject) => {
-    callIpc('renderer:mount-collection', { collectionUid, collectionPathname, brunoConfig })
-      .then(() => dispatch(updateCollectionMountStatus({ collectionUid, mountStatus: 'mounted' })))
-      .then(resolve)
-      .catch(() => {
-        dispatch(updateCollectionMountStatus({ collectionUid, mountStatus: 'unmounted' }));
-        reject();
-      });
+    const { ipcRenderer } = window;
+    ipcRenderer.invoke('renderer:load-large-request', { collectionUid, pathname }).then(resolve).catch(reject);
   });
 };
 
-  export const showInFolder = (collectionPath) => () => {
-    return new Promise((resolve, reject) => {
-      const { ipcRenderer } = window;
-      ipcRenderer.invoke('renderer:show-in-folder', collectionPath).then(resolve).catch(reject);
+export const mountCollection =
+  ({ collectionUid, collectionPathname, brunoConfig }) =>
+  (dispatch, getState) => {
+    dispatch(updateCollectionMountStatus({ collectionUid, mountStatus: 'mounting' }));
+    return new Promise(async (resolve, reject) => {
+      callIpc('renderer:mount-collection', { collectionUid, collectionPathname, brunoConfig })
+        .then(() => dispatch(updateCollectionMountStatus({ collectionUid, mountStatus: 'mounted' })))
+        .then(resolve)
+        .catch(() => {
+          dispatch(updateCollectionMountStatus({ collectionUid, mountStatus: 'unmounted' }));
+          reject();
+        });
     });
   };
 
-export const updateRunnerConfiguration = (collectionUid, selectedRequestItems, requestItemsOrder) => (dispatch) => {
-  dispatch(_updateRunnerConfiguration({
-    collectionUid,
-    selectedRequestItems,
-    requestItemsOrder
-  }));
+export const showInFolder = (collectionPath) => () => {
+  return new Promise((resolve, reject) => {
+    const { ipcRenderer } = window;
+    ipcRenderer.invoke('renderer:show-in-folder', collectionPath).then(resolve).catch(reject);
+  });
 };
+
+export const updateRunnerConfiguration = (collectionUid, selectedRequestItems, requestItemsOrder, delay) => (dispatch) => {
+  dispatch(
+    _updateRunnerConfiguration({
+      collectionUid,
+      selectedRequestItems,
+      requestItemsOrder,
+    delay
+    })
+  );
+};
+
+export const updateActiveConnectionsInStore = (activeConnectionIds) => (dispatch, getState) => {
+  dispatch(updateActiveConnections(activeConnectionIds));
+};
+
+export const openCollectionSettings =
+  (collectionUid, tabName = 'overview') =>
+  (dispatch, getState) => {
+    const state = getState();
+    const collection = findCollectionByUid(state.collections.collections, collectionUid);
+
+    return new Promise((resolve, reject) => {
+      if (!collection) {
+        return reject(new Error('Collection not found'));
+      }
+
+      dispatch(
+        updateSettingsSelectedTab({
+          collectionUid: collection.uid,
+          tab: tabName
+        })
+      );
+
+      dispatch(
+        addTab({
+          uid: collection.uid,
+          collectionUid: collection.uid,
+          type: 'collection-settings'
+        })
+      );
+
+      resolve();
+    });
+  };
