@@ -23,10 +23,47 @@ import mime from 'mime-types';
 import path from 'utils/common/path';
 import { getUniqueTagsFromItems } from 'utils/collections/index';
 
+// gRPC status code meanings
+const grpcStatusCodes = {
+  0: 'OK',
+  1: 'CANCELLED',
+  2: 'UNKNOWN',
+  3: 'INVALID_ARGUMENT',
+  4: 'DEADLINE_EXCEEDED',
+  5: 'NOT_FOUND',
+  6: 'ALREADY_EXISTS',
+  7: 'PERMISSION_DENIED',
+  8: 'RESOURCE_EXHAUSTED',
+  9: 'FAILED_PRECONDITION',
+  10: 'ABORTED',
+  11: 'OUT_OF_RANGE',
+  12: 'UNIMPLEMENTED',
+  13: 'INTERNAL',
+  14: 'UNAVAILABLE',
+  15: 'DATA_LOSS',
+  16: 'UNAUTHENTICATED'
+};
+
 const initialState = {
   collections: [],
-  collectionSortOrder: 'default'
+  collectionSortOrder: 'default',
+  activeConnections: []
 };
+
+const initiatedGrpcResponse = {
+  statusCode: null,
+  statusText: 'STREAMING',
+  statusDescription: null,
+  headers: [],
+  metadata: null,
+  trailers: null,
+  statusDetails: null,
+  error: null,
+  isError: false,
+  duration: 0,
+  responses: [],
+  timestamp: Date.now(),
+}
 
 export const collectionsSlice = createSlice({
   name: 'collections',
@@ -325,6 +362,167 @@ export const collectionsSlice = createSlice({
           });
         }
       }
+    },
+    runGrpcRequestEvent: (state, action) => {
+      const { itemUid, collectionUid, eventType, eventData } = action.payload;
+      const collection = findCollectionByUid(state.collections, collectionUid);
+      if (!collection) return;
+      
+      const item = findItemInCollection(collection, itemUid);
+      if (!item) return;
+      const request = item.draft ? item.draft.request : item.request;
+      const isUnary = request.methodType === 'unary';
+
+      if (eventType === 'request') {
+        item.requestSent = eventData;
+        item.requestSent.timestamp = Date.now();
+        item.response = {
+          initiatedGrpcResponse,
+          statusText: isUnary ? 'PENDING' : 'STREAMING'
+        };
+      }
+
+      if (!collection.timeline) {
+        collection.timeline = [];
+      }
+
+      collection.timeline.push({
+        type: "request",
+        eventType: eventType, // Add the specific gRPC event type
+        collectionUid: collection.uid,
+        folderUid: null,
+        itemUid: item.uid,
+        timestamp: Date.now(),
+        data: {
+          request: eventData || item.requestSent || item.request,
+          timestamp: Date.now(),
+          eventData: eventData,
+        }
+      });
+      
+    },
+    grpcResponseReceived: (state, action) => {
+      const { itemUid, collectionUid, eventType, eventData } = action.payload;
+      const collection = findCollectionByUid(state.collections, collectionUid);
+    
+      if (!collection) return;
+
+      const item = findItemInCollection(collection, itemUid);
+
+      if (!item) return;
+      
+      // Get current response state or create initial state
+      const currentResponse = item.response || initiatedGrpcResponse
+      const timestamp = item?.requestSent?.timestamp;
+      let updatedResponse = { ...currentResponse, duration: Date.now() - (timestamp || Date.now()) };
+
+      
+      // Process based on event type
+      switch (eventType) {
+        case 'response':
+          const { error, res } = eventData;
+          
+          //  Handle error if present
+          if (error) {
+            const errorCode = error.code || 2; // Default to UNKNOWN if no code
+            
+            updatedResponse.error = error.details || 'gRPC error occurred';
+            updatedResponse.statusCode = errorCode;
+            updatedResponse.statusText = grpcStatusCodes[errorCode] || 'UNKNOWN';
+            updatedResponse.errorDetails = error;
+            updatedResponse.isError = true;
+          }
+
+          // Add response to list
+          updatedResponse.responses = res 
+            ? [...(currentResponse?.responses || []), res] 
+            : [...(currentResponse?.responses || [])];
+          break;
+          
+        case 'metadata':
+          updatedResponse.headers = eventData.metadata;
+          updatedResponse.metadata = eventData.metadata;
+          break;
+          
+        case 'status':
+          // Extract status info
+          const statusCode = eventData.status?.code;
+          const statusDetails = eventData.status?.details;
+          const statusMetadata = eventData.status?.metadata;
+          
+          // Set status based on actual code and details
+          updatedResponse.statusCode = statusCode;
+          updatedResponse.statusText = grpcStatusCodes[statusCode] || 'UNKNOWN';
+          updatedResponse.statusDescription = statusDetails;
+          updatedResponse.statusDetails = eventData.status;
+          
+          // Store trailers (status metadata)
+          if (statusMetadata) {
+            updatedResponse.trailers = statusMetadata;
+          }
+          
+          // Handle error status (non-zero code)
+          if (statusCode !== 0) {
+            updatedResponse.isError = true;
+            updatedResponse.error = statusDetails || `gRPC error with code ${statusCode} (${updatedResponse.statusText})`;
+          }
+          
+          break;
+          
+        case 'error':
+          // Extract error details
+          const errorCode = eventData.error?.code || 2; // Default to UNKNOWN if no code
+          const errorDetails = eventData.error?.details || eventData.error?.message;
+          const errorMetadata = eventData.error?.metadata;
+          
+          updatedResponse.isError = true;
+          updatedResponse.error = errorDetails || 'Unknown gRPC error';
+          updatedResponse.statusCode = errorCode;
+          updatedResponse.statusText = grpcStatusCodes[errorCode] || 'UNKNOWN';
+          updatedResponse.statusDescription = errorDetails;
+          
+          // Store error metadata as trailers if present
+          if (errorMetadata) {
+            updatedResponse.trailers = errorMetadata;
+          }
+          
+          break;
+          
+        case 'end':
+          state.activeConnections = state.activeConnections.filter(id => id !== itemUid);
+          break;
+          
+        case 'cancel':
+          updatedResponse.statusCode = 1; // CANCELLED
+          updatedResponse.statusText = 'CANCELLED';
+          updatedResponse.statusDescription = 'Stream cancelled by client or server';
+          state.activeConnections = state.activeConnections.filter(id => id !== itemUid);
+          break;
+      }
+      
+      item.requestState = 'received';
+      item.response = updatedResponse;
+
+      // Update the timeline
+      if (!collection?.timeline) {
+        collection.timeline = [];
+      }
+
+      // Append the new timeline entry with specific gRPC event type
+      collection.timeline.push({
+        type: "request",
+        eventType: eventType, // Add the specific gRPC event type
+        collectionUid: collection.uid,
+        folderUid: null,
+        itemUid: item.uid,
+        timestamp: Date.now(),
+        data: {
+          request: item.requestSent || item.request,
+          response: updatedResponse,
+          eventData: eventData, // Store the original event data
+          timestamp: Date.now(),
+        }
+      });
     },
     responseCleared: (state, action) => {
       const collection = findCollectionByUid(state.collections, action.payload.collectionUid);
@@ -1183,6 +1381,7 @@ export const collectionsSlice = createSlice({
           if (!item.draft) {
             item.draft = cloneDeep(item);
           }
+          
           switch (item.draft.request.body.mode) {
             case 'json': {
               item.draft.request.body.json = action.payload.content;
@@ -1210,6 +1409,10 @@ export const collectionsSlice = createSlice({
             }
             case 'multipartForm': {
               item.draft.request.body.multipartForm = action.payload.content;
+              break;
+            }
+            case 'grpc': {
+              item.draft.request.body.grpc = action.payload.content;
               break;
             }
           }
@@ -1303,6 +1506,21 @@ export const collectionsSlice = createSlice({
             item.draft = cloneDeep(item);
           }
           item.draft.request.method = action.payload.method;
+          item.draft.request.methodType = action.payload.methodType;
+        }
+      }
+    },
+    updateRequestProtoPath: (state, action) => {
+      const collection = findCollectionByUid(state.collections, action.payload.collectionUid);
+
+      if (collection) {
+        const item = findItemInCollection(collection, action.payload.itemUid);    
+
+        if (item && isItemARequest(item)) {
+          if (!item.draft) {
+            item.draft = cloneDeep(item);
+          }
+          item.draft.request.protoPath = action.payload.protoPath;
         }
       }
     },
@@ -2291,12 +2509,13 @@ export const collectionsSlice = createSlice({
       }
     },
     updateRunnerConfiguration: (state, action) => {
-      const { collectionUid, selectedRequestItems, requestItemsOrder } = action.payload;
+      const { collectionUid, selectedRequestItems, requestItemsOrder, delay } = action.payload;
       const collection = findCollectionByUid(state.collections, collectionUid);
       if (collection) {
         collection.runnerConfiguration = {
           selectedRequestItems: selectedRequestItems || [],
-          requestItemsOrder: requestItemsOrder || []
+          requestItemsOrder: requestItemsOrder || [],
+          delay: delay
         };
       }
     },
@@ -2460,7 +2679,10 @@ export const collectionsSlice = createSlice({
       if (collection) {
         collection.allTags = getUniqueTagsFromItems(collection.items);
       }
-    }
+    },
+    updateActiveConnections: (state, action) => {
+      state.activeConnections = [...action.payload.activeConnectionIds];
+    },
   }
 });
 
@@ -2487,6 +2709,8 @@ export const {
   processEnvUpdateEvent,
   requestCancelled,
   responseReceived,
+  runGrpcRequestEvent,
+  grpcResponseReceived,
   responseCleared,
   clearTimeline,
   clearRequestTimeline,
@@ -2531,6 +2755,7 @@ export const {
   updateResponseScript,
   updateRequestTests,
   updateRequestMethod,
+  updateRequestProtoPath,
   addAssertion,
   updateAssertion,
   deleteAssertion,
@@ -2584,7 +2809,8 @@ export const {
   updateFolderAuthMode,
   addRequestTag,
   deleteRequestTag,
-  updateCollectionTagsList
+  updateCollectionTagsList,
+  updateActiveConnections
 } = collectionsSlice.actions;
 
 export default collectionsSlice.reducer;
