@@ -60,7 +60,8 @@ const STATIC_API_HINTS = {
     'bru.getEnvVar(key)',
     'bru.getFolderVar(key)',
     'bru.getCollectionVar(key)',
-    'bru.setEnvVar(key,value)',
+    'bru.setEnvVar(key, value)',
+    'bru.setEnvVar(key, value, options)',
     'bru.deleteEnvVar(key)',
     'bru.hasVar(key)',
     'bru.getVar(key)',
