@@ -20,6 +20,7 @@ const { setBrunoConfig } = require('../store/bruno-config');
 const EnvironmentSecretsStore = require('../store/env-secrets');
 const UiStateSnapshot = require('../store/ui-state-snapshot');
 const { parseBruFileMeta, hydrateRequestWithUuid } = require('../utils/collection');
+const { parseLargeRequestWithRedaction } = require('../utils/parse');
 
 const MAX_FILE_SIZE = 2.5 * 1024 * 1024;
 
@@ -344,11 +345,17 @@ const addDirectory = async (win, pathname, collectionUid, collectionPath) => {
   let seq;
   const folderBruFilePath = path.join(pathname, `folder.bru`);
 
-  if (fs.existsSync(folderBruFilePath)) {
-    let folderBruFileContent = fs.readFileSync(folderBruFilePath, 'utf8');
-    let folderBruData = await parseFolder(folderBruFileContent);
-    name = folderBruData?.meta?.name || name;
-    seq = folderBruData?.meta?.seq;
+  try {
+    if (fs.existsSync(folderBruFilePath)) {
+      let folderBruFileContent = fs.readFileSync(folderBruFilePath, 'utf8');
+      let folderBruData = await parseFolder(folderBruFileContent);
+      name = folderBruData?.meta?.name || name;
+      seq = folderBruData?.meta?.seq;
+    }
+  }
+  catch(error) {
+    console.error('Error occured while parsing folder.bru file!');
+    console.error(error);
   }
 
   const directory = {
@@ -462,10 +469,20 @@ const change = async (win, pathname, collectionUid, collectionPath) => {
       };
 
       const bru = fs.readFileSync(pathname, 'utf8');
-      file.data = await parseRequest(bru);
+      const fileStats = fs.statSync(pathname);
 
-      hydrateRequestWithUuid(file.data, pathname);
-      win.webContents.send('main:collection-tree-updated', 'change', file);
+      if (fileStats.size >= MAX_FILE_SIZE) {
+        const parsedData = await parseLargeRequestWithRedaction(bru);
+        file.data = parsedData;
+        file.size = sizeInMB(fileStats?.size);
+        hydrateRequestWithUuid(file.data, pathname);
+        win.webContents.send('main:collection-tree-updated', 'change', file);
+      } else {
+        file.data = await parseRequest(bru);
+        file.size = sizeInMB(fileStats?.size);
+        hydrateRequestWithUuid(file.data, pathname);
+        win.webContents.send('main:collection-tree-updated', 'change', file);
+      }
     } catch (err) {
       console.error(err);
     }
