@@ -148,6 +148,7 @@ const getOAuth2TokenUsingAuthorizationCode = async ({ request, collectionUid, fo
     credentialsId,
     autoRefreshToken,
     autoFetchToken,
+    additionalParameters,
   } = oAuth;
   const url = requestCopy?.oauth2?.accessTokenUrl;
   
@@ -242,7 +243,7 @@ const getOAuth2TokenUsingAuthorizationCode = async ({ request, collectionUid, fo
   }
 
   // Fetch new token process
-  const { authorizationCode, debugInfo } = await getOAuth2AuthorizationCode(requestCopy, codeChallenge, collectionUid);
+  let { authorizationCode, debugInfo } = await getOAuth2AuthorizationCode(requestCopy, codeChallenge, collectionUid);
 
   let axiosRequestConfig = {};
   axiosRequestConfig.method = 'POST';
@@ -251,26 +252,30 @@ const getOAuth2TokenUsingAuthorizationCode = async ({ request, collectionUid, fo
     'Accept': 'application/json',
   };
   if (credentialsPlacement === "basic_auth_header") {
-    axiosRequestConfig.headers['Authorization'] = `Basic ${Buffer.from(`${clientId}:${clientSecret}`).toString('base64')}`;
+    axiosRequestConfig.headers['Authorization'] = `Basic ${Buffer.from(`${encodeURIComponent(clientId)}:${encodeURIComponent(clientSecret)}`).toString('base64')}`;
   }
   const data = {
     grant_type: 'authorization_code',
     code: authorizationCode,
     redirect_uri: callbackUrl,
-    client_id: clientId,
   };
-  if (clientSecret && credentialsPlacement !== "basic_auth_header") {
+  if (credentialsPlacement !== "basic_auth_header") {
+    data.client_id = clientId;
+  }
+  if (clientSecret && clientSecret.trim() !== '' && credentialsPlacement !== "basic_auth_header") {
     data.client_secret = clientSecret;
   }
   if (pkce) {
     data['code_verifier'] = codeVerifier;
   }
-  if (scope && scope.trim() !== '') {
-    data.scope = scope;
-  }
-  axiosRequestConfig.data = qs.stringify(data);
+
   axiosRequestConfig.url = url;
   axiosRequestConfig.responseType = 'arraybuffer';
+  // Apply additional parameters to token request
+  if (additionalParameters?.token?.length) {
+    applyAdditionalParameters(axiosRequestConfig, data, additionalParameters.token);
+  }
+  axiosRequestConfig.data = qs.stringify(data);
   try {
     const { credentials, requestDetails } = await getCredentialsFromTokenUrl({ requestConfig: axiosRequestConfig, certsAndProxyConfig });
 
@@ -292,7 +297,7 @@ const getOAuth2TokenUsingAuthorizationCode = async ({ request, collectionUid, fo
 const getOAuth2AuthorizationCode = (request, codeChallenge, collectionUid) => {
   return new Promise(async (resolve, reject) => {
     const { oauth2 } = request;
-    const { callbackUrl, clientId, authorizationUrl, scope, state, pkce, accessTokenUrl } = oauth2;
+    const { callbackUrl, clientId, authorizationUrl, scope, state, pkce, accessTokenUrl, additionalParameters } = oauth2;
 
     const authorizationUrlWithQueryParams = new URL(authorizationUrl);
     authorizationUrlWithQueryParams.searchParams.append('response_type', 'code');
@@ -310,18 +315,44 @@ const getOAuth2AuthorizationCode = (request, codeChallenge, collectionUid) => {
     if (state) {
       authorizationUrlWithQueryParams.searchParams.append('state', state);
     }
+    if (additionalParameters?.authorization?.length) {
+      additionalParameters.authorization.forEach(param => {
+        if (param.enabled && param.name) {
+          if (param.sendIn === 'queryparams') {
+            authorizationUrlWithQueryParams.searchParams.append(param.name, param.value || '');
+          }
+        }
+      });
+    }
+    
     try {
       const authorizeUrl = authorizationUrlWithQueryParams.toString();
       const { authorizationCode, debugInfo } = await authorizeUserInWindow({
         authorizeUrl,
         callbackUrl,
-        session: oauth2Store.getSessionIdOfCollection({ collectionUid, url: accessTokenUrl })
+        session: oauth2Store.getSessionIdOfCollection({ collectionUid, url: accessTokenUrl }),
+        additionalHeaders: getAdditionalHeaders(additionalParameters?.authorization)
       });
       resolve({ authorizationCode, debugInfo });
     } catch (err) {
       reject(err);
     }
   });
+};
+
+const getAdditionalHeaders = (params) => {
+  if (!params || !params.length) {
+    return {};
+  }
+  
+  const headers = {};
+  params.forEach(param => {
+    if (param.enabled && param.name && param.sendIn === 'headers') {
+      headers[param.name] = param.value || '';
+    }
+  });
+  
+  return headers;
 };
 
 // CLIENT CREDENTIALS
@@ -337,6 +368,7 @@ const getOAuth2TokenUsingClientCredentials = async ({ request, collectionUid, fo
     credentialsId,
     autoRefreshToken,
     autoFetchToken,
+    additionalParameters,
   } = oAuth;
 
   const url = requestCopy?.oauth2?.accessTokenUrl;
@@ -354,15 +386,6 @@ const getOAuth2TokenUsingClientCredentials = async ({ request, collectionUid, fo
   if (!clientId) {
     return {
       error: 'Client ID is required for OAuth2 client credentials flow',
-      credentials: null,
-      url,
-      credentialsId
-    };
-  }
-
-  if (!clientSecret) {
-    return {
-      error: 'Client Secret is required for OAuth2 client credentials flow',
       credentials: null,
       url,
       credentialsId
@@ -427,22 +450,27 @@ const getOAuth2TokenUsingClientCredentials = async ({ request, collectionUid, fo
     'content-type': 'application/x-www-form-urlencoded',
     'Accept': 'application/json',
   };
-  if (credentialsPlacement === "basic_auth_header") {
-    axiosRequestConfig.headers['Authorization'] = `Basic ${Buffer.from(`${clientId}:${clientSecret}`).toString('base64')}`;
+  if (credentialsPlacement === "basic_auth_header" && clientSecret && clientSecret.trim() !== '') {
+    axiosRequestConfig.headers['Authorization'] = `Basic ${Buffer.from(`${encodeURIComponent(clientId)}:${encodeURIComponent(clientSecret)}`).toString('base64')}`;
   }
   const data = {
     grant_type: 'client_credentials',
-    client_id: clientId,
   };
-  if (clientSecret && credentialsPlacement !== "basic_auth_header") {
+  if (credentialsPlacement !== "basic_auth_header") {
+    data.client_id = clientId;
+  }
+  if (clientSecret && clientSecret.trim() !== '' && credentialsPlacement !== "basic_auth_header") {
     data.client_secret = clientSecret;
   }
   if (scope && scope.trim() !== '') {
     data.scope = scope;
   }
-  axiosRequestConfig.data = qs.stringify(data);
   axiosRequestConfig.url = url;
   axiosRequestConfig.responseType = 'arraybuffer';
+  if (additionalParameters?.token?.length) {
+    applyAdditionalParameters(axiosRequestConfig, data, additionalParameters.token);
+  }
+  axiosRequestConfig.data = qs.stringify(data);
   let debugInfo = { data: [] };
   try {
     const { credentials, requestDetails } = await getCredentialsFromTokenUrl({ requestConfig: axiosRequestConfig, certsAndProxyConfig });
@@ -469,6 +497,7 @@ const getOAuth2TokenUsingPasswordCredentials = async ({ request, collectionUid, 
     credentialsId,
     autoRefreshToken,
     autoFetchToken,
+    additionalParameters,
   } = oAuth;
   const url = requestCopy?.oauth2?.accessTokenUrl;
 
@@ -568,24 +597,29 @@ const getOAuth2TokenUsingPasswordCredentials = async ({ request, collectionUid, 
     'content-type': 'application/x-www-form-urlencoded',
     'Accept': 'application/json',
   };
-  if (credentialsPlacement === "basic_auth_header") {
-    axiosRequestConfig.headers['Authorization'] = `Basic ${Buffer.from(`${clientId}:${clientSecret}`).toString('base64')}`;
+  if (credentialsPlacement === "basic_auth_header" && clientSecret && clientSecret.trim() !== '') {
+    axiosRequestConfig.headers['Authorization'] = `Basic ${Buffer.from(`${encodeURIComponent(clientId)}:${encodeURIComponent(clientSecret)}`).toString('base64')}`;
   }
   const data = {
     grant_type: 'password',
     username,
     password,
-    client_id: clientId,
   };
-  if (clientSecret && credentialsPlacement !== "basic_auth_header") {
+  if (credentialsPlacement !== "basic_auth_header") {
+    data.client_id = clientId;
+  }
+  if (clientSecret && clientSecret.trim() !== '' && credentialsPlacement !== "basic_auth_header") {
     data.client_secret = clientSecret;
   }
   if (scope && scope.trim() !== '') {
     data.scope = scope;
   }
-  axiosRequestConfig.data = qs.stringify(data);
   axiosRequestConfig.url = url;
   axiosRequestConfig.responseType = 'arraybuffer';
+  if (additionalParameters?.token?.length) {
+    applyAdditionalParameters(axiosRequestConfig, data, additionalParameters.token);
+  }
+  axiosRequestConfig.data = qs.stringify(data);
   let debugInfo = { data: [] };
   try {
     const { credentials, requestDetails } = await getCredentialsFromTokenUrl({ requestConfig: axiosRequestConfig, certsAndProxyConfig });
@@ -599,7 +633,7 @@ const getOAuth2TokenUsingPasswordCredentials = async ({ request, collectionUid, 
 
 const refreshOauth2Token = async ({ requestCopy, collectionUid, certsAndProxyConfig }) => {
   const oAuth = get(requestCopy, 'oauth2', {});
-  const { clientId, clientSecret, credentialsId } = oAuth;
+  const { clientId, clientSecret, credentialsId, credentialsPlacement, additionalParameters } = oAuth;
   const url = oAuth.refreshTokenUrl ? oAuth.refreshTokenUrl : oAuth.accessTokenUrl;
 
   const credentials = getStoredOauth2Credentials({ collectionUid, url, credentialsId });
@@ -610,10 +644,12 @@ const refreshOauth2Token = async ({ requestCopy, collectionUid, certsAndProxyCon
   } else {
     const data = {
       grant_type: 'refresh_token',
-      client_id: clientId,
       refresh_token: credentials.refresh_token,
     };
-    if (clientSecret) {
+    if (credentialsPlacement !== "basic_auth_header") {
+      data.client_id = clientId;
+    }
+    if (clientSecret && clientSecret.trim() !== '' && credentialsPlacement !== "basic_auth_header") {
       data.client_secret = clientSecret;
     }
     let axiosRequestConfig = {};
@@ -622,9 +658,15 @@ const refreshOauth2Token = async ({ requestCopy, collectionUid, certsAndProxyCon
       'content-type': 'application/x-www-form-urlencoded',
       'Accept': 'application/json'
     };
-    axiosRequestConfig.data = qs.stringify(data);
+    if (credentialsPlacement === "basic_auth_header") {
+      axiosRequestConfig.headers['Authorization'] = `Basic ${Buffer.from(`${encodeURIComponent(clientId)}:${encodeURIComponent(clientSecret)}`).toString('base64')}`;
+    }
     axiosRequestConfig.url = url;
     axiosRequestConfig.responseType = 'arraybuffer';
+    if (additionalParameters?.refresh?.length) {
+      applyAdditionalParameters(axiosRequestConfig, data, additionalParameters.refresh);
+    }
+    axiosRequestConfig.data = qs.stringify(data);
     let debugInfo = { data: [] };
     try {
       const { credentials, requestDetails } = await getCredentialsFromTokenUrl({ requestConfig: axiosRequestConfig, certsAndProxyConfig });
@@ -659,6 +701,36 @@ const generateCodeChallenge = (codeVerifier) => {
   return base64Hash;
 };
 
+// Apply additional parameters to a request
+const applyAdditionalParameters = (requestCopy, data, params = []) => {
+  params.forEach(param => {
+    if (!param.enabled || !param.name) {
+      return;
+    }
+
+    switch (param.sendIn) {
+      case 'headers':
+        requestCopy.headers[param.name] = param.value || '';
+        break;
+      case 'queryparams':
+        // For query params, add to URL
+        try {
+          let url = new URL(requestCopy.url);
+          url.searchParams.append(param.name, param.value || '');
+          requestCopy.url = url.href;
+        }
+        catch (error) {
+          console.error('invalid token/refresh url', requestCopy.url);
+        }
+        break;
+      case 'body':
+        // For body, add to data object
+        data[param.name] = param.value || '';
+        break;
+    }
+  });
+}
+
 const getOAuth2TokenUsingImplicitGrant = async ({ request, collectionUid, forceFetch = false }) => {
   const { oauth2 = {} } = request;
   const {
@@ -668,7 +740,8 @@ const getOAuth2TokenUsingImplicitGrant = async ({ request, collectionUid, forceF
     state = '',
     callbackUrl,
     credentialsId = 'credentials',
-    autoFetchToken = true
+    autoFetchToken = true,
+    additionalParameters
   } = oauth2;
 
   // Validate required fields
@@ -753,6 +826,15 @@ const getOAuth2TokenUsingImplicitGrant = async ({ request, collectionUid, forceF
   if (state) {
     authorizationUrlWithQueryParams.searchParams.append('state', state);
   }
+  if (additionalParameters?.authorization?.length) {
+    additionalParameters.authorization.forEach(param => {
+      if (param.enabled && param.name) {
+        if (param.sendIn === 'queryparams') {
+          authorizationUrlWithQueryParams.searchParams.append(param.name, param.value || '');
+        }
+      }
+    });
+  }
 
   const authorizeUrl = authorizationUrlWithQueryParams.toString();
   
@@ -761,7 +843,8 @@ const getOAuth2TokenUsingImplicitGrant = async ({ request, collectionUid, forceF
       authorizeUrl,
       callbackUrl,
       session: oauth2Store.getSessionIdOfCollection({ collectionUid, url: authorizationUrl }),
-      grantType: 'implicit'
+      grantType: 'implicit',
+      additionalHeaders: getAdditionalHeaders(additionalParameters?.authorization)
     });
 
     if (!implicitTokens || !implicitTokens.access_token) {
