@@ -3,7 +3,7 @@ module.exports = {
     '^.+\\.(ts|js)$': 'babel-jest',
   },
   transformIgnorePatterns: [
-    '/node_modules/(?!(lodash-es)/)',
+    '/node_modules/(?!(lodash-es|is-ip|ip-regex|super-regex|function-timeout|time-span|convert-hrtime|clone-regexp|is-regexp)/)'
   ],
   testEnvironment: 'node',
   testMatch: [
