@@ -11,12 +11,13 @@ export const test = baseTest.extend<
     page: Page;
     newPage: Page;
     pageWithUserData: Page;
+    restartApp: (options?: { initUserDataPath?: string }) => Promise<ElectronApplication>;
   },
   {
     createTmpDir: (tag?: string) => Promise<string>;
-    launchElectronApp: (options?: { initUserDataPath?: string }) => Promise<ElectronApplication>;
+    launchElectronApp: (options?: { initUserDataPath?: string; userDataPath?: string }) => Promise<ElectronApplication>;
     electronApp: ElectronApplication;
-    reuseOrLaunchElectronApp: (options?: { initUserDataPath?: string }) => Promise<ElectronApplication>;
+    reuseOrLaunchElectronApp: (options?: { initUserDataPath?: string; userDataPath?: string }) => Promise<ElectronApplication>;
   }
 >({
   createTmpDir: [
@@ -37,8 +38,13 @@ export const test = baseTest.extend<
   launchElectronApp: [
     async ({ playwright, createTmpDir }, use, workerInfo) => {
       const apps: ElectronApplication[] = [];
-      await use(async ({ initUserDataPath } = {}) => {
-        const userDataPath = await createTmpDir('electron-userdata');
+      await use(async ({ initUserDataPath, userDataPath: providedUserDataPath } = {}) => {
+        const userDataPath = providedUserDataPath || (await createTmpDir('electron-userdata'));
+
+        // Ensure dir exists when caller supplies their own path
+        if (providedUserDataPath) {
+          await fs.promises.mkdir(userDataPath, { recursive: true });
+        }
 
         if (initUserDataPath) {
           const replacements = {
@@ -62,17 +68,22 @@ export const test = baseTest.extend<
           args: [electronAppPath],
           env: {
             ...process.env,
-            ELECTRON_USER_DATA_PATH: userDataPath,
+            ELECTRON_USER_DATA_PATH: userDataPath
           }
         });
 
         const { workerIndex } = workerInfo;
-        app.process().stdout.on('data', (data) => {
-          process.stdout.write(data.toString().replace(/^(?=.)/gm, `[Electron #${workerIndex}] |`));
-        });
-        app.process().stderr.on('data', (error) => {
-          process.stderr.write(error.toString().replace(/^(?=.)/gm, `[Electron #${workerIndex}] |`));
-        });
+        const electronProcess = app.process();
+        if (electronProcess?.stdout) {
+          electronProcess.stdout.on('data', (data) => {
+            process.stdout.write(data.toString().replace(/^(?=.)/gm, `[Electron #${workerIndex}] |`));
+          });
+        }
+        if (electronProcess?.stderr) {
+          electronProcess.stderr.on('data', (error) => {
+            process.stderr.write(error.toString().replace(/^(?=.)/gm, `[Electron #${workerIndex}] |`));
+          });
+        }
 
         apps.push(app);
         return app;
@@ -137,18 +148,46 @@ export const test = baseTest.extend<
   reuseOrLaunchElectronApp: [
     async ({ launchElectronApp }, use, testInfo) => {
       const apps: Record<string, ElectronApplication> = {};
-      await use(async ({ initUserDataPath } = {}) => {
-        const key = initUserDataPath;
+      await use(async ({ initUserDataPath, userDataPath } = {}) => {
+        const key = userDataPath || initUserDataPath;
         if (key && apps[key]) {
           return apps[key];
         }
-        const app = await launchElectronApp({ initUserDataPath });
-        apps[key] = app;
+        const app = await launchElectronApp({ initUserDataPath, userDataPath });
+        if (key) {
+          apps[key] = app;
+        }
         return app;
       });
     },
     { scope: 'worker' }
   ],
+
+  restartApp: async ({ launchElectronApp }, use, testInfo) => {
+    const appInstances: Array<{ app: ElectronApplication; initUserDataPath?: string }> = [];
+    await use(async ({ initUserDataPath } = {}) => {
+      // Get the test directory and check for init-user-data folder
+      const testDir = path.dirname(testInfo.file);
+      const defaultInitUserDataPath = path.join(testDir, 'init-user-data');
+
+      // Use provided initUserDataPath, or check if default path exists, or use undefined
+      let userDataPath = initUserDataPath;
+      if (!userDataPath) {
+        const hasInitUserData = await fs.promises.stat(defaultInitUserDataPath).catch(() => false);
+        userDataPath = hasInitUserData ? defaultInitUserDataPath : undefined;
+      }
+
+      const app = await launchElectronApp({ initUserDataPath: userDataPath });
+      appInstances.push({ app, initUserDataPath: userDataPath });
+      return app;
+    });
+
+    // Clean up all app instances
+    for (const { app } of appInstances) {
+      await app.context().close();
+      await app.close();
+    }
+  },
 
   pageWithUserData: async ({ reuseOrLaunchElectronApp }, use, testInfo) => {
     const testDir = path.dirname(testInfo.file);
